doctr/ticketocr/requirements.txt
Django==4.2.7
Pillow==10.0.1
pytesseract==0.3.10
doctr==0.7.0
torch==2.1.0
torchvision==0.16.0
pdf2image==1.16.3
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.5
google-generativeai==0.3.2
docling==1.0.0
openpyxl==3.1.2
